Fix violation of Fortitude rule PORT011
Checks for use of raw number literals as kinds

Rather than setting an intrinsic type's kind using an integer literal, such as
`real(8)` or `integer(kind=4)`, consider setting kinds using parameters in the
intrinsic module `iso_fortran_env` such as `real64` and `int32`. For
C-compatible types, consider instead `iso_c_binding` types such as
`real(c_double)`.

Although it is widely believed that `real(8)` represents an 8-byte floating
point (and indeed, this is the case for most compilers and architectures),
there is nothing in the standard to mandate this, and compiler vendors are free
to choose any mapping between kind numbers and machine precision. This may lead
to surprising results if your code is ported to another machine or compiler.

For floating point variables, we recommended using `real(sp)` (single
precision), `real(dp)` (double precision), and `real(qp)` (quadruple precision),
using:

```f90
use, intrinsic :: iso_fortran_env, only: sp => real32, &
                                         dp => real64, &
                                         qp => real128
```

Or alternatively:

```f90
integer, parameter :: sp = selected_real_kind(6, 37)
integer, parameter :: dp = selected_real_kind(15, 307)
integer, parameter :: qp = selected_real_kind(33, 4931)
```

Some prefer to set one precision parameter `wp` (working precision), which is
set in one module and used throughout a project.

Integer sizes may be set similarly:

```f90
integer, parameter :: i1 = selected_int_kind(2)  ! 8 bits
integer, parameter :: i2 = selected_int_kind(4)  ! 16 bits
integer, parameter :: i4 = selected_int_kind(9)  ! 32 bits
integer, parameter :: i8 = selected_int_kind(18) ! 64 bits
```

Or:

```f90
use, intrinsic :: iso_fortran_env, only: i1 => int8, &
                                         i2 => int16, &
                                         i4 => int32, &
                                         i8 => int64
```

+++ src/trans/gpu/algor/ext_acc.F90
@@ -46,9 +46,10 @@ module openacc_ext
 contains
 
   function ext_acc_pass_2d_r4(arr) result(ret)
+    use iso_fortran_env, only: real32
     implicit none
     type(ext_acc_arr_desc) :: ret
-    real(4), intent(in), target :: arr(:,:)
+    real(kind=real32), intent(in), target :: arr(:,:)
 
     type(c_ptr) :: ptr1, ptr2
     integer(c_size_t) :: ptr1_v, ptr2_v
@@ -70,9 +71,10 @@ contains
     ret%sz = ret%sz + (ptr2_v - ptr1_v) * size(arr, 1)
   end function
   function ext_acc_pass_3d_r4(arr) result(ret)
+    use iso_fortran_env, only: real32
     implicit none
     type(ext_acc_arr_desc) :: ret
-    real(4), intent(in), target :: arr(:,:,:)
+    real(kind=real32), intent(in), target :: arr(:,:,:)
 
     type(c_ptr) :: ptr1, ptr2
     integer(c_size_t) :: ptr1_v, ptr2_v
@@ -94,9 +96,10 @@ contains
     ret%sz = ret%sz + (ptr2_v - ptr1_v) * size(arr, 2)
   end function
   function ext_acc_pass_4d_r4(arr) result(ret)
+    use iso_fortran_env, only: real32
     implicit none
     type(ext_acc_arr_desc) :: ret
-    real(4), intent(in), target :: arr(:,:,:,:)
+    real(kind=real32), intent(in), target :: arr(:,:,:,:)
 
     type(c_ptr) :: ptr1, ptr2
     integer(c_size_t) :: ptr1_v, ptr2_v
@@ -118,9 +121,10 @@ contains
     ret%sz = ret%sz + (ptr2_v - ptr1_v) * size(arr, 3)
   end function
   function ext_acc_pass_2d_r8(arr) result(ret)
+    use iso_fortran_env, only: real64
     implicit none
     type(ext_acc_arr_desc) :: ret
-    real(8), intent(in), target :: arr(:,:)
+    real(kind=real64), intent(in), target :: arr(:,:)
 
     type(c_ptr) :: ptr1, ptr2
     integer(c_size_t) :: ptr1_v, ptr2_v
@@ -142,9 +146,10 @@ contains
     ret%sz = ret%sz + (ptr2_v - ptr1_v) * size(arr, 1)
   end function
   function ext_acc_pass_3d_r8(arr) result(ret)
+    use iso_fortran_env, only: real64
     implicit none
     type(ext_acc_arr_desc) :: ret
-    real(8), intent(in), target :: arr(:,:,:)
+    real(kind=real64), intent(in), target :: arr(:,:,:)
 
     type(c_ptr) :: ptr1, ptr2
     integer(c_size_t) :: ptr1_v, ptr2_v
@@ -166,9 +171,10 @@ contains
     ret%sz = ret%sz + (ptr2_v - ptr1_v) * size(arr, 2)
   end function
   function ext_acc_pass_4d_r8(arr) result(ret)
+    use iso_fortran_env, only: real64
     implicit none
     type(ext_acc_arr_desc) :: ret
-    real(8), intent(in), target :: arr(:,:,:,:)
+    real(kind=real64), intent(in), target :: arr(:,:,:,:)
 
     type(c_ptr) :: ptr1, ptr2
     integer(c_size_t) :: ptr1_v, ptr2_v
@@ -266,7 +272,7 @@ contains
     integer(acc_handle_kind), optional :: stream
 #endif
 #ifdef OMPGPU
-    integer(4), optional :: stream
+    integer(kind=int32), optional :: stream
 #endif
 
     type(common_pointer_descr), allocatable :: common_ptrs(:)
@@ -298,19 +304,20 @@ contains
 #ifdef ACCGPU
     use openacc, only: acc_async_sync
 #endif
+    use iso_fortran_env, only: int32
     implicit none
     type(ext_acc_arr_desc), intent(in) :: ptrs(:)
 #ifdef ACCGPU
     integer(acc_handle_kind), optional :: stream
 #endif
 #ifdef OMPGPU
-    integer(4), optional :: stream
+    integer(kind=int32), optional :: stream
 #endif
 
     type(common_pointer_descr), allocatable :: common_ptrs(:)
 
     integer :: i, num_ranges
-    integer(4), pointer :: pp(:)
+    integer(kind=int32), pointer :: pp(:)
 #ifdef ACCGPU
     integer(acc_handle_kind) :: stream_act
     if (present(stream)) then
@@ -336,18 +343,19 @@ contains
 #ifdef ACCGPU
     use openacc, only: acc_async_sync
 #endif
+    use iso_fortran_env, only: int32
     implicit none
     type(ext_acc_arr_desc), intent(in) :: ptrs(:)
 #ifdef ACCGPU
     integer(acc_handle_kind), optional :: stream
 #endif
 #ifdef OMPGPU
-    integer(4), optional :: stream
+    integer(kind=int32), optional :: stream
 #endif
     type(common_pointer_descr), allocatable :: common_ptrs(:)
 
     integer :: i, num_ranges
-    integer(4), pointer :: pp(:)
+    integer(kind=int32), pointer :: pp(:)
 #ifdef ACCGPU
     integer(acc_handle_kind) :: stream_act
     if (present(stream)) then
@@ -373,17 +381,18 @@ contains
 #ifdef ACCGPU
     use openacc, only: acc_async_sync
 #endif
+    use iso_fortran_env, only: int32
     implicit none
     type(ext_acc_arr_desc), intent(in) :: ptrs(:)
 #ifdef ACCGPU
     integer(acc_handle_kind), optional :: stream
 #else
-    integer(4), optional :: stream
+    integer(kind=int32), optional :: stream
 #endif
     type(common_pointer_descr), allocatable :: common_ptrs(:)
 
     integer :: i, num_ranges
-    integer(4), pointer :: pp(:)
+    integer(kind=int32), pointer :: pp(:)
 #ifdef ACCGPU
     integer(acc_handle_kind) :: stream_act
